requirements.txt
pandas>=2.0
tqdm>=4.65
psycopg2-binary>=2.9

sentence-transformers>=2.2
torch>=2.0
transformers>=4.35

fastapi
uvicorn
transformers
sentence-transformers
torch
pydantic
pillow
